requirements.txt
streamlit==1.32.0
nibabel==5.1.0
numpy==1.24.0
torch==2.0.0
scipy==1.10.0
scikit-learn==1.2.0
matplotlib==3.7.0
Pillow==10.0.0